Migrate regressions from 3501-4000
